+++ ContentPatcher/Framework/Patches/EditDataPatchField.cs
@@ -74,5 +74,11 @@ namespace ContentPatcher.Framework.Patches
         {
             return this.Contextuals.GetTokensUsed();
         }
+
+        /// <summary>Get diagnostic info about the contextual instance.</summary>
+        public IContextualState GetDiagnosticState()
+        {
+            return this.Contextuals.GetDiagnosticState();
+        }
     }
 }