+++ ContentPatcher/Framework/Patches/EditDataPatchMoveRecord.cs
@@ -77,5 +77,11 @@ namespace ContentPatcher.Framework.Patches
         {
             return this.Contextuals.GetTokensUsed();
         }
+
+        /// <summary>Get diagnostic info about the contextual instance.</summary>
+        public IContextualState GetDiagnosticState()
+        {
+            return this.Contextuals.GetDiagnosticState();
+        }
     }
 }